backend/scheduler/requirements.txt
APScheduler==3.11.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
